Rename all instances of column to field
