--- ModelDefinition/Column/Boolean.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace ModelDefinition\Column;
-
-use ModelDefinition\Column\Interfaces\ColumnInterface;
-
-final class Boolean extends AbstractColumn implements ColumnInterface
-{
-}
