--- ModelDefinition/Column/Integer.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace ModelDefinition\Column;
-
-use ModelDefinition\Column\AbstractColumn;
-use ModelDefinition\Column\Interfaces\ColumnInterface;
-
-final class Integer extends AbstractColumn implements ColumnInterface
-{
-}
