--- ModelDefinition/Column/Interfaces/ColumnInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace ModelDefinition\Column\Interfaces;
-
-interface ColumnInterface
-{
-    public function name() : string;
-}
