--- ModelDefinition/Column/Text.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace ModelDefinition\Column;
-
-use ModelDefinition\Column\Interfaces\ColumnInterface;
-
-final class Text extends AbstractColumn implements ColumnInterface
-{
-}
