+++ ModelDefinition/Field/Boolean.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ModelDefinition\Field;
+
+use ModelDefinition\Field\Interfaces\FieldInterface;
+
+final class Boolean extends AbstractField implements FieldInterface
+{
+}
