+++ ModelDefinition/Field/Integer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ModelDefinition\Field;
+
+use ModelDefinition\Field\AbstractField;
+use ModelDefinition\Field\Interfaces\FieldInterface;
+
+final class Integer extends AbstractField implements FieldInterface
+{
+}
