+++ ModelDefinition/Field/Interfaces/FieldInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ModelDefinition\Field\Interfaces;
+
+interface FieldInterface
+{
+    public function name() : string;
+}
