+++ ModelDefinition/Field/Text.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ModelDefinition\Field;
+
+use ModelDefinition\Field\Interfaces\FieldInterface;
+
+final class Text extends AbstractField implements FieldInterface
+{
+}
